Import example of generating Jira issues using template & csv
This example uses the 'go-jira' app to create Jira issues,
based on a template file (used by go-jira) and a CSV file
(used to generate the issue files).

I'm building a go-jira container in this example because the
last release of go-jira was two years ago, but it still gets
fixes in the mainline branch.